@@ -2,10 +2,7 @@ name: shared_preferences
 description: Flutter plugin for reading and writing simple key-value pairs.
   Wraps NSUserDefaults on iOS and SharedPreferences on Android.
 homepage: https://github.com/flutter/plugins/tree/master/packages/shared_preferences/shared_preferences
-# 0.5.y+z is compatible with 1.0.0, if you land a breaking change bump
-# the version to 2.0.0.
-# See more details: https://github.com/flutter/flutter/wiki/Package-migration-to-1.0.0
-version: 0.5.13+2
+version: 2.0.0-nullsafety.1
 
 flutter:
   plugin:
@@ -26,16 +23,16 @@ dependencies:
   meta: ^1.0.4
   flutter:
     sdk: flutter
-  shared_preferences_platform_interface: ^1.0.0
+  shared_preferences_platform_interface: ^2.0.0-nullsafety
   # The design on https://flutter.dev/go/federated-plugins was to leave
   # this constraint as "any". We cannot do it right now as it fails pub publish
   # validation, so we set a ^ constraint.
   # TODO(franciscojma): Revisit this (either update this part in the  design or the pub tool).
   # https://github.com/flutter/flutter/issues/46264
-  shared_preferences_linux: ^0.0.2
-  shared_preferences_macos: ^0.0.1
-  shared_preferences_web: ^0.1.2
-  shared_preferences_windows: ^0.0.1
+  shared_preferences_linux: ^0.0.4-nullsafety
+  shared_preferences_macos: ^0.0.2-nullsafety
+  shared_preferences_web: ^0.2.0-nullsafety
+  shared_preferences_windows: ^0.0.3-nullsafety
 
 dev_dependencies:
   flutter_test:
@@ -47,5 +44,5 @@ dev_dependencies:
   pedantic: ^1.8.0
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: ">=1.12.13+hotfix.5"
