@@ -4,8 +4,13 @@ description: Demonstrates how to use the shared_preferences plugin.
 dependencies:
   flutter:
     sdk: flutter
-  shared_preferences: any
+  shared_preferences_platform_interface: ^2.0.0-nullsafety
   shared_preferences_macos:
+    # When depending on this package from a real application you should use:
+    #   shared_preferences_macos: ^x.y.z
+    # See https://dart.dev/tools/pub/dependencies#version-constraints
+    # The example app is bundled with the plugin so we use a path dependency on
+    # the parent directory to use the current plugin's version.
     path: ../
 
 dev_dependencies:
@@ -19,5 +24,5 @@ flutter:
   uses-material-design: true
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.12.8 <2.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
+  flutter: ">=1.12.8"
