@@ -2,7 +2,8 @@ name: shared_preferences_windows_example
 description: Demonstrates how to use the shared_preferences_windows plugin.
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
+  flutter: ">=1.12.8"
 
 dependencies:
   flutter:
@@ -11,12 +12,18 @@ dependencies:
 
 dependency_overrides:
   shared_preferences_windows:
+    # When depending on this package from a real application you should use:
+    #   shared_preferences_windows: ^x.y.z
+    # See https://dart.dev/tools/pub/dependencies#version-constraints
+    # The example app is bundled with the plugin so we use a path dependency on
+    # the parent directory to use the current plugin's version.
     path: ../
 
 dev_dependencies:
   flutter_driver:
     sdk: flutter
-  e2e: ^0.2.0
+  integration_test:
+    path: ../../../integration_test
   pedantic: ^1.8.0
 
 flutter:
