@@ -44,6 +44,8 @@ abstract class CameraPlatform extends PlatformInterface {
   }
 
   /// Completes with a list of available cameras.
+  ///
+  /// This method returns an empty list when no cameras are available.
   Future<List<CameraDescription>> availableCameras() {
     throw UnimplementedError('availableCameras() is not implemented.');
   }
@@ -52,7 +54,7 @@ abstract class CameraPlatform extends PlatformInterface {
   Future<int> createCamera(
     CameraDescription cameraDescription,
     ResolutionPreset? resolutionPreset, {
-    bool enableAudio = true,
+    bool enableAudio = false,
   }) {
     throw UnimplementedError('createCamera() is not implemented.');
   }
