@@ -1,4 +1,4 @@
-// Autogenerated from Pigeon (v0.1.12), do not edit directly.
+// Autogenerated from Pigeon (v0.1.19), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 
 package io.flutter.plugins.videoplayer;
